Info updated.

+++ stall/forms/submit.py
@@ -3,7 +3,7 @@ from stall.forms.bases import BaseForm
 
 
 class SubmitForm(BaseForm):
-    booth = forms.IntegerField(required=False)
+    booth = forms.FloatField(required=False)
     remarks = forms.CharField(widget=forms.Textarea, required=False)
     agreement = forms.BooleanField()
     number_of_people = forms.IntegerField(required=False)
